requirements.txt
pandas
numpy
scikit-learn
joblib
dvc
streamlit>=1.30.0
altair>=5
playwright
python-dotenv